arm64: dts: qcom: msm8917: fix cpu idle states
Like on MSM8953, we need to use force-psci-domains on this SoCs.
Also, we cannot just do cluster power collapse like that (it freezes):
since MSM8917 is a single-cluster SoC, cluster PC is system PC.

pwrdn/ghds (4) is the only cluster state we can use.
retention (2) seems broken (hangs), looks like that's why it was marked
as worse than ghds in the downstream dts.

Signed-off-by: Val Packett <val@packett.cool>