MSM8996 net: qrtr: Turn QRTR into a bus
Implement a QRTR bus to allow for creating drivers for individual QRTR
services. With this in place, devices are dynamically registered for QRTR
services as they become available, and drivers for these devices are
matched using service and instance IDs.

In smd.c, replace all current occurences of qdev with qsdev in order to
distinguish between the newly added QRTR device which represents a QRTR
service with the existing QRTR SMD device which represents the endpoint
through which services are provided.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>